Add support of publish/consume with job attributes (#227)
This PR introduces the new API: PublishJob and RePublishJob to allow
the use of the job attributes and mark the old API as deprecated. But we
don't add the new API for the batch publish since we think it's not a rigorous
implementation and don't encourage users to use that.